backend/requirements.txt
Flask>=2.0.0
flask-cors>=3.0.10
pymongo>=3.11.0 
python-dotenv>=0.19.0 
python-dateutil>=2.8.0 